+++ betterrandom/src/main/java/io/github/pr0methean/betterrandom/prng/package-info.java
@@ -1,4 +1,7 @@
 /**
  * Subclasses of {@link java.util.Random} using different random-number generation algorithms.
  */
+@ParametersAreNonnullByDefault
 package io.github.pr0methean.betterrandom.prng;
+
+import javax.annotation.ParametersAreNonnullByDefault;