+++ betterrandom/src/main/java/io/github/pr0methean/betterrandom/seed/package-info.java
@@ -3,4 +3,7 @@
  * especially those that also implement
  * {@link io.github.pr0methean.betterrandom.ByteArrayReseedableRandom}.
  */
+@ParametersAreNonnullByDefault
 package io.github.pr0methean.betterrandom.seed;
+
+import javax.annotation.ParametersAreNonnullByDefault;