+++ betterrandom/src/test/java/io/github/pr0methean/betterrandom/seed/SeedGeneratorTest.java
@@ -2,6 +2,8 @@ package io.github.pr0methean.betterrandom.seed;
 
 import static org.testng.Assert.assertEquals;
 
+import nl.jqno.equalsverifier.EqualsVerifier;
+import nl.jqno.equalsverifier.api.SingleTypeEqualsVerifierApi;
 import org.powermock.api.mockito.mockpolicies.Slf4jMockPolicy;
 import org.powermock.core.classloader.annotations.MockPolicy;
 import org.powermock.modules.testng.PowerMockTestCase;
@@ -28,6 +30,14 @@ public abstract class SeedGeneratorTest<T extends SeedGenerator> extends PowerMo
   @Test public void testToString() {
     Assert.assertNotNull(seedGenerator.toString(),
         "toString() returned null for a " + seedGenerator.getClass().getSimpleName());
+  }
+
+  @Test public void testWithEqualsVerifier() {
+    createEqualsVerifier().verify();
+  }
+
+  protected SingleTypeEqualsVerifierApi<? extends SeedGenerator> createEqualsVerifier() {
+    return EqualsVerifier.forClass(seedGenerator.getClass()).usingGetClass();
   }
 
   /**